diy_small_erp/requirements.txt
Django==2.2
django-crispy-forms==1.9.2
django-registration-redux==2.8
django-simple-history==2.12.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
